Simplify HBChannel inheritance
